requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai==0.3.1
python-dotenv==1.0.0
langchain>=0.0.310
langchain-community>=0.0.10
langchain-core>=0.1.8
chromadb==0.4.15
pandas==2.1.0
gunicorn==21.2.0
